requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
amqp==5.0.6
arrow==1.1.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
blessed==1.18.0
ccxt==1.50.4
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
Django==3.2.3
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-picklefield==3.0.1
django-q==1.3.6
djangorestframework==3.12.4
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.0.1
kombu==5.0.2
multidict==5.1.0
prompt-toolkit==3.0.18
psycopg2==2.7.7
pycares==4.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.4.1
